Umar | get expense by user id middleware

+++ server/controllers/expense.controller.js
@@ -1,4 +1,4 @@
-import Expense from "../models/expense.model.js";
+import Expense from "../models/Expense.Model.js";
 import errorHandler from "../errors/error.js";
 
 export const getExpense = async (req, res, next) => {
@@ -50,13 +50,34 @@ export const updateExpense = async (req, res, next) => {
 export const deleteExpense = async (req, res, next) => {
   try {
     const { expenseId } = req.params;
+    const { expensesIds } = req;
+
+    // handle case where the param expenseId is not a valid id
+    // handle case whre the requested expenseId does not even exists
+
+    // Check if expenseId exists in userExpenses
+    let expenseToDelete = false;
+    for (const _expenseId of expensesIds) {
+      if (_expenseId.toString() === expenseId) {
+        expenseToDelete = true;
+        break;
+      }
+    }
+
+    // throw error in case expense not found
+    if (!expenseToDelete) {
+      return next(
+        errorHandler(403, "You are not allowed to delete this expense")
+      );
+    }
+
+    // Proceed with deleting the expense
     const deletedExpense = await Expense.findByIdAndDelete(expenseId);
     if (!deletedExpense) {
       return next(errorHandler(404, "Expense not found"));
     }
-    res
-      .status(200)
-      .send({ success: true, message: "Expense deleted successfully" });
+
+    res.status(200).send({ success: true, message: "Expense deleted" });
   } catch (error) {
     next(error);
   }
